backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
PyMuPDF
langchain
psycopg2-binary
python-dotenv